feat: rename remove plural
