+++ apps/backend/src/resolvers/mutations/ApproveVisitRegistrationMutation.ts
